changes in approprice file structure
